+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+PaymentMethodAvailability.swift
@@ -45,6 +45,8 @@ extension PaymentSheet {
         .mbWay,
         .bizum,
         .kakaoPay,
+        .naverPay,
+        .krCard,
         .payco,
         .sequra,
     ]
@@ -362,11 +364,11 @@ extension PaymentSheet {
 
     /// Payment method types that require mandate data for PaymentIntents when `setup_future_usage` is set
     static var requiresMandateDataForPaymentIntent: Set<STPPaymentMethodType> {
-        [.alipay, .payPal, .cashApp, .revolutPay, .amazonPay, .klarna, .satispay, .twint, .kakaoPay]
+        [.alipay, .payPal, .cashApp, .revolutPay, .amazonPay, .klarna, .satispay, .twint, .kakaoPay, .naverPay, .krCard]
     }
 
     /// Payment method types that require mandate data for SetupIntents
     static var requiresMandateDataForSetupIntent: Set<STPPaymentMethodType> {
-        [.alipay, .payPal, .revolutPay, .satispay, .twint, .kakaoPay]
+        [.alipay, .payPal, .revolutPay, .satispay, .twint, .kakaoPay, .naverPay, .krCard]
     }
 }