+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/ExpectedFormHierarchies.swift
@@ -52,6 +52,36 @@ enum ExpectedFormHierarchy {
         }
     }
 
+    // MARK: - Naver Pay
+
+    enum NaverPay {
+        static var paymentIntent: FormHierarchyNode {
+            FormHierarchyNode(type: "FormElement", children: [
+                FormHierarchyNode(type: "SectionElement", children: [
+                    FormHierarchyNode(type: "DropdownFieldElement", properties: ["itemCount": "2", "label": "Select how you want to pay:"])
+                ]),
+            ])
+        }
+        static var settingUp: FormHierarchyNode {
+            FormHierarchyNode(type: "FormElement", children: [
+                FormHierarchyNode(type: "SectionElement", children: [
+                    FormHierarchyNode(type: "DropdownFieldElement", properties: ["itemCount": "2", "label": "Select how you want to pay:"])
+                ]),
+                FormHierarchyNode(type: "SimpleMandateElement", properties: ["text": "By confirming your payment with NICEPAY, you allow..."]),
+            ])
+        }
+    }
+
+    // MARK: - Korean cards
+
+    enum KoreanCards {
+        static var paymentIntent: FormHierarchyNode { emptyForm }
+        static var settingUp: FormHierarchyNode {
+            FormHierarchyNode(type: "FormElement", children: [
+                FormHierarchyNode(type: "SimpleMandateElement", properties: ["text": "By confirming your payment with NICEPAY, you allow..."])
+            ])
+        }
+    }
     // MARK: - Alipay
 
     enum Alipay {