+++ ci_scripts/l10n/config.rb
@@ -7,6 +7,7 @@ DEVELOPMENT_LANGUAGE = 'en'
 
 # Languages that we localize to.
 LANGUAGES = %w[
+  ar-SA
   bg-BG
   ca-ES
   cs-CZ